# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 学脉 | UniPulse Asia API
  version: 1.0.0
  description: Academic social network connecting students across Asia-Pacific universities
paths:
...
  /api/communities/:
    get:
      operationId: api_communities_list
      description: |-
        社区API
        支持筛选参数：category, city, oncampus, study_group
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - api
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommunityList'
          description: ''
    post:
      operationId: api_communities_create
      description: |-
        社区API
        支持筛选参数：category, city, oncampus, study_group
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Community'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Community'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Community'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
  /api/communities/{slug}/:
    get:
      operationId: api_communities_retrieve
      description: |-
        社区API
        支持筛选参数：category, city, oncampus, study_group
      parameters:
      - in: path
        name: slug
        schema:
          type: string
          title: URL别名
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    put:
      operationId: api_communities_update
      description: |-
        社区API
        支持筛选参数：category, city, oncampus, study_group
      parameters:
      - in: path
        name: slug
        schema:
          type: string
          title: URL别名
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Community'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Community'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Community'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    patch:
      operationId: api_communities_partial_update
      description: |-
        社区API
        支持筛选参数：category, city, oncampus, study_group
      parameters:
      - in: path
        name: slug
        schema:
          type: string
          title: URL别名
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    delete:
      operationId: api_communities_destroy
      description: |-
        社区API
        支持筛选参数：category, city, oncampus, study_group
      parameters:
      - in: path
        name: slug
        schema:
          type: string
          title: URL别名
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/notifications/{id}/mark_as_read/:
    post:
      operationId: api_notifications_mark_as_read_create
      description: 标记单条通知为已读
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Notification'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Notification'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
          description: ''
  /api/notifications/clear_all/:
    delete:
      operationId: api_notifications_clear_all_destroy
      description: 清空当前用户的所有通知
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/notifications/mark_all_as_read/:
    post:
      operationId: api_notifications_mark_all_as_read_create
      description: 标记所有通知为已读（别名，向后兼容）
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Notification'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Notification'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
          description: ''
...
  /api/notifications/unread_count/:
    get:
      operationId: api_notifications_unread_count_retrieve
      description: 获取未读通知数量（按类型统计）
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
          description: ''
...
components:
  schemas:
...
    CategoryEnum:
      enum:
      - interest
      - city
      - oncampus
      - study_group
      type: string
      description: |-
        * `interest` - 兴趣爱好
        * `city` - 城市
        * `oncampus` - 校园
        * `study_group` - 学习小组
...
    Community:
      type: object
      description: Community serializer
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: 社区名称
          maxLength: 200
        slug:
          type: string
          title: URL别名
          maxLength: 220
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          title: 描述
        cover_url:
          type: string
          format: uri
          nullable: true
          title: 封面图片URL
          maxLength: 200
        category:
          allOf:
          - $ref: '#/components/schemas/CategoryEnum'
          title: 分类
        city:
          type: string
          nullable: true
          title: 城市
          maxLength: 100
        is_oncampus:
          type: boolean
          title: 是否校园社区
        is_study_group:
          type: boolean
          title: 是否学习小组
        members:
          type: integer
          readOnly: true
          title: 成员数
        activity_rate:
          type: number
          format: double
          title: 活跃度 0-1
        visibility:
          $ref: '#/components/schemas/Visibility1fdEnum'
        is_published:
          type: boolean
        joined:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - joined
      - members
      - name
      - slug
      - updated_at
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        notification_type:
          type: string
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/NotificationTypeEnum'
          readOnly: true
          title: 类型
        title:
          type: string
          readOnly: true
          title: 标题
        content:
          type: string
          readOnly: true
          title: 内容
        message:
          type: string
          readOnly: true
        sender:
          allOf:
          - $ref: '#/components/schemas/Sender'
          readOnly: true
        related_post:
          allOf:
          - $ref: '#/components/schemas/RelatedPost'
          readOnly: true
        related_comment:
          allOf:
          - $ref: '#/components/schemas/RelatedComment'
          readOnly: true
        link:
          type: string
          readOnly: true
          title: 链接
        is_read:
          type: boolean
          title: 已读
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
        read_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: 阅读时间
        time_ago:
          type: string
          readOnly: true
      required:
      - content
      - created_at
      - id
      - link
      - message
      - notification_type
      - read_at
      - related_comment
      - related_post
      - sender
      - time_ago
      - title
      - type
    NotificationTypeEnum:
      enum:
      - like
      - comment
      - follow
      - mention
      - system
      type: string
      description: |-
        * `like` - Like
        * `comment` - Comment
        * `follow` - Follow
        * `mention` - Mention
        * `system` - System
...
    PaginatedCommunityList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Community'
...
    PatchedCommunity:
      type: object
      description: Community serializer
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: 社区名称
          maxLength: 200
        slug:
          type: string
          title: URL别名
          maxLength: 220
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          title: 描述
        cover_url:
          type: string
          format: uri
          nullable: true
          title: 封面图片URL
          maxLength: 200
        category:
          allOf:
          - $ref: '#/components/schemas/CategoryEnum'
          title: 分类
        city:
          type: string
          nullable: true
          title: 城市
          maxLength: 100
        is_oncampus:
          type: boolean
          title: 是否校园社区
        is_study_group:
          type: boolean
          title: 是否学习小组
        members:
          type: integer
          readOnly: true
          title: 成员数
        activity_rate:
          type: number
          format: double
          title: 活跃度 0-1
        visibility:
          $ref: '#/components/schemas/Visibility1fdEnum'
        is_published:
          type: boolean
        joined:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    RelatedComment:
      type: object
      description: 关联评论信息
      properties:
        id:
          type: integer
          readOnly: true
        content:
          type: string
          title: 评论内容
      required:
      - content
      - id
    RelatedPost:
      type: object
      description: 关联帖子信息
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          title: 标题
          maxLength: 200
      required:
      - id
...
    Sender:
      type: object
      description: 发送者信息
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          title: 用户名
          description: 必填；长度为150个字符或以下；只能包含字母、数字、特殊字符“@”、“.”、“-”和“_”。
          pattern: ^[\w.@+-]+$
          maxLength: 150
        full_name:
          type: string
          readOnly: true
        avatar:
          type: string
          readOnly: true
      required:
      - avatar
      - full_name
      - id
      - username
...
    Visibility1fdEnum:
      enum:
      - public
      - university
      - private
      type: string
      description: |-
        * `public` - Public
        * `university` - University
        * `private` - Private
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT